Refactoring Major restructuring to Position library
