+++ src/dataStructures/Player.ts
@@ -1,9 +1,8 @@
-import { Settings } from "../settings.js"
 import GenericFactory from "./genericObject.js"
 
-export default function PlayerFactory(location: TVector, velocity: TVector, rotation: Degrees): IRotatableGeneric & IAcceleration {
+export default radius => (location, velocity, rotation) => {
   return {
-    ...GenericFactory(location, velocity, Settings.PLAYER_RADIUS, ObjectType.Player),
+    ...GenericFactory(location, velocity, radius, ObjectType.Player),
     rotation,
     acceleration: .05,
   }
